+++ src/keys.rs
@@ -38,6 +38,17 @@ pub fn generate_seed(
       full_entropy = wallet.seed_secret.full_entropy.clone();
       mnemonic_dictionary = wallet.seed_secret.mnemonic_dictionary.clone();
     }
+    "QRNG" => {
+      let raw_entropy: Zeroizing<String> = wallet.seed_secret.raw_entropy.clone();
+
+      let entropy_checksum: Zeroizing<String> = e_q::calculate_checksum_for_entropy(raw_entropy.clone());
+      wallet.seed_secret.entropy_checksum = entropy_checksum.clone();
+
+      full_entropy = Zeroizing::new(format!("{}{}", *raw_entropy, *entropy_checksum));
+      wallet.seed_secret.full_entropy = full_entropy.clone();
+
+      mnemonic_dictionary = wallet.seed_secret.mnemonic_dictionary.clone();
+    }
     "RNG" => {
       // TODO: Implement RNG Mnemonic passphrase + add to GUI
       let entropy_length: Zeroizing<usize> = wallet.seed_secret.entropy_length.clone();