+++ src/main.rs
@@ -238,6 +238,7 @@ struct GuiSettings {
   secrets_dialog: crypt::ShowSecretsDialog,
   anu_dialog: crypt::ShowAnuDialog,
 
+  hide_private_keys: bool,
   unify_evm: bool,
   unify_master_keys: bool,
   hardened_address: bool,
@@ -261,6 +262,7 @@ impl GuiSettings {
       secrets_dialog: crypt::ShowSecretsDialog::new(),
       anu_dialog: crypt::ShowAnuDialog::new(),
 
+      hide_private_keys: true,
       unify_evm: false,
       unify_master_keys: true,
       hardened_address: true,
@@ -317,8 +319,8 @@ impl EgoQuantum {
       _ => self.get_bip(),
     };
 
-    if self.wallet.seed_secret.seed.is_empty() {
-      match keys::generate_seed(&mut self.wallet, entropy_source) {
+    if self.wallet.seed_secret.full_entropy.is_empty() {
+      match keys::generate_seed(&mut self.wallet, entropy_source.clone()) {
         Ok(_) => {}
         Err(err) => {
           return Err(AppError::log(format!("Problem with generating seed: {}", err)));
@@ -347,9 +349,21 @@ impl EgoQuantum {
     let active_coins = if cfg!(feature = "dev") { 2 } else { 1 };
 
     // TODO: Add address_count as GUI parameters
-    let source = self.get_entropy_source();
-    let index = if *source == "SVG" { *self.wallet.address_components.derivation_path.last_index } else { 0 };
-    let address_count = std::cmp::max(self.gui.address_count, index);
+    let address_count = 10;
+    let last_index = *self.wallet.address_components.derivation_path.last_index;
+
+    let (start_index, end_index) = if entropy_source.as_str() == "SVG" {
+      if self.wallet.addresses_by_coin.0.is_empty() {
+        // Bootstrap SVG mode
+        (0, last_index)
+      } else {
+        // Continue paging
+        (last_index, last_index.saturating_add(address_count))
+      }
+    } else {
+      // Normal mode
+      (last_index, last_index.saturating_add(address_count))
+    };
 
     // ECDB: Extended Coin DataBase
     let resource_path = std::path::Path::new("coin").join("ECDB.csv");
@@ -384,7 +398,7 @@ impl EgoQuantum {
             self.wallet.address_components.wallet_import_format = Zeroizing::new(columns[10].to_string());
             self.wallet.address_components.evm = Zeroizing::new(columns[11].trim().eq_ignore_ascii_case("true"));
 
-            for address_index in 0..address_count {
+            for address_index in start_index..end_index {
               self.wallet.address_components.derivation_path.address = Zeroizing::new(address_index);
 
               match self.wallet.address_components.key_derivation.as_str() {
@@ -432,7 +446,7 @@ impl EgoQuantum {
         }
       }
 
-      *self.wallet.address_components.derivation_path.last_index += address_count;
+      *self.wallet.address_components.derivation_path.last_index = end_index;
     }
 
     Ok(())
@@ -466,21 +480,35 @@ impl EgoQuantum {
           )
           .on_hover_text_at_pointer(descriptions[0]);
 
-          // QRNG
-          #[cfg(feature = "dev")]
-          {
-            let resp = ui
-              .selectable_value(
-                &mut self.wallet.seed_secret.entropy_source,
-                Zeroizing::new(VALID_ENTROPY_SOURCES[1].to_string()),
-                VALID_ENTROPY_SOURCES[1],
-              )
-              .on_hover_text_at_pointer(descriptions[1]);
+          // // QRNG
+          ui.selectable_value(
+            &mut self.wallet.seed_secret.entropy_source,
+            Zeroizing::new(VALID_ENTROPY_SOURCES[1].to_string()),
+            VALID_ENTROPY_SOURCES[1],
+          )
+          .on_hover_text_at_pointer(descriptions[1]);
 
-            if resp.clicked() {
-              self.gui.anu_dialog.open = true;
-            }
-          }
+          // #[cfg(feature = "dev")]
+          // {
+          //   let resp = ui
+          //     .selectable_value(
+          //       &mut self.wallet.seed_secret.entropy_source,
+          //       Zeroizing::new(VALID_ENTROPY_SOURCES[1].to_string()),
+          //       VALID_ENTROPY_SOURCES[1],
+          //     )
+          //     .on_hover_text_at_pointer(descriptions[1]);
+
+          //   if resp.clicked() {
+          //     self.gui.anu_dialog.open = true;
+
+          //     if !self.gui.anu_dialog.randomized_entropy.is_empty() && !self.gui.anu_dialog.open {
+          //       let _ = self.generate_new_wallet(
+          //         Some(Zeroizing::new(VALID_ENTROPY_SOURCES[1].to_string())),
+          //         Some(self.gui.anu_dialog.randomized_entropy.clone()),
+          //       );
+          //     }
+          //   }
+          // }
 
           // FILE
           #[cfg(feature = "dev")]
@@ -694,6 +722,18 @@ impl EgoQuantum {
       });
 
       ui.menu_button("Privacy", |ui| {
+        let hide_private_keys_label = [
+          "When enabled:",
+          "Private keys will be hidden until you move mouse over it.",
+          "\n",
+          "When disabled:",
+          "All private keys will be visible.",
+        ];
+
+        let hide_private_keys_resp: egui::Response = ui.add_enabled(true,egui::Checkbox::new(&mut self.gui.hide_private_keys, "Hide private keys"));
+        hide_private_keys_resp.on_hover_text(hide_private_keys_label.join("\n")).on_disabled_hover_text(&devel);
+
+
         let evm_label = [
           "When enabled:",
           "Normalize how EVM addresses are displayed so they look the same across all networks. This improves usability when managing multiple chains.",
@@ -908,7 +948,11 @@ impl EgoQuantum {
                     ui.ctx().copy_text(first.private_key.to_string());
                   }
 
-                  let display_text = if ui.ui_contains_pointer() { &first.private_key } else { "••••••••••••••••" };
+                  let display_text = if ui.ui_contains_pointer() {
+                    &first.private_key
+                  } else {
+                    if self.gui.hide_private_keys { "••••••••••••••••" } else { &first.private_key }
+                  };
                   ui.label(display_text);
                 });
               });
@@ -960,8 +1004,11 @@ impl EgoQuantum {
                         ui.ctx().copy_text(addr.private_key.to_string());
                       }
 
-                      let display_text =
-                        if ui.ui_contains_pointer() { &addr.private_key } else { "••••••••••••••••" };
+                      let display_text = if ui.ui_contains_pointer() {
+                        &addr.private_key
+                      } else {
+                        if self.gui.hide_private_keys { "••••••••••••••••" } else { &addr.private_key }
+                      };
                       ui.label(display_text);
                     });
                   });
@@ -995,7 +1042,22 @@ impl EgoQuantum {
       if self.wallet.addresses_by_coin.0.len() < self.gui.max_rows {
         if ui.button(button_label).clicked() {
           let source = self.get_entropy_source();
-          let _ = self.generate_new_wallet(Some(source));
+
+          match source.as_str() {
+            "QRNG" => {
+              self.gui.anu_dialog.open = true;
+
+              if self.gui.anu_dialog.save_entropy {
+                self.wallet.seed_secret.raw_entropy = self.gui.anu_dialog.randomized_entropy.clone();
+
+                let _ = self.generate_new_wallet(Some(source));
+              }
+            }
+            _ => {
+              // TODO: Improve
+              let _ = self.generate_new_wallet(Some(source));
+            }
+          }
         }
       } else {
         ui.label("Memory limit reached—cannot generate more addresses.");